Begin moving state & messages out of Timeline/Posts into MessageList
